feat: Add `DestinasiEdit` navigation destination
Creates a new `DestinasiEdit` object to define the navigation route for the item editing screen.

This object implements the `DestinasiNavigasi` interface and specifies the route `item_edit` and the title resource `R.string.edit_siswa`. It also defines a route with an argument, `idSiswa`, for navigating to a specific item.